requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
aiohttp==3.9.1
python-dotenv==1.0.0
